refactor(test-harness): one type per file — split the per-suite statistics
stats.rs held eight structs. Each moves to its own module under
output/summary/ (its lowest common parent), with its tests: DiffCounter,
Movement, TopMovers, PairedBytes, CompileAggregate, FailureRegression,
FailureRegressions, and SuiteStats. The rendering impls on these types stay in
render.rs for now; they follow in the render split.
